Recast lighting.
