+++ raytracer/src/lib.rs
@@ -1,0 +1,7 @@
+pub mod camera;
+pub mod config;
+pub mod materials;
+pub mod point3d;
+pub mod ray;
+pub mod raytracer;
+pub mod sphere;
